+++ src/main/java/com/github/jlangch/venice/impl/functions/CoreFunctions.java
@@ -729,7 +729,7 @@ public class CoreFunctions {
 				VncFunction
 					.meta()
 					.arglists("(== x y)")
-					.doc("Returns true if both operands have the equivalent type")
+					.doc("Returns true if both operands have the equivalent type and value")
 					.examples("(== 0 0)", "(== 0 1)", "(== 0 0.0)")
 					.build()
 		) {
@@ -748,7 +748,7 @@ public class CoreFunctions {
 				VncFunction
 					.meta()
 					.arglists("(!= x y)")
-					.doc("Returns true if both operands do not have the equivalent type")
+					.doc("Returns true if both operands do not have the equivalent type and value")
 					.examples("(!= 0 1)", "(!= 0 0)", "(!= 0 0.0)")
 					.build()
 		) {