+++ ChangeLog.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 
+- Java «raw string literals» style triple quoted strings
 - function `=`. Returns true if both operands have equivalent type and value