+++ ChangeLog.md
@@ -12,7 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 
 - Java «raw string literals» style triple quoted strings
-- function `=`. Returns true if both operands have equivalent type and value
+- function `=`. Returns true if both operands have equivalent type and value. `(= 1 1.0) -> false`, `(== 1 1.0) -> true`